Multi-LLM-NER-KPI-with-Backend/requirements.txt
# Python dependencies for Multi-LLM NER Backend

# Web framework
Flask==3.0.0
flask-cors==4.0.0

# AI/LLM SDKs
anthropic==0.20.0
openai==1.12.0
google-generativeai==0.4.0
groq==0.4.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Development
pytest==8.0.0
pytest-flask==1.3.0
